1933 Fixed links in routers (#1972)
* Create LoginPickerPage

Create LoginPickerPage. This page give user two choices to log in, samf3 and samf4.

* Update loginPickerPage with biome fix

* Update samf-three.ts

* Biome fix and stylelint checks

Biome fix --write and stylechecks fixes from last tuesday that didnt get pushed.

* Update fix to links in router

Update fix to a link in LoginPickerPage that was by accident passed as a String and not an object.

* Fixed format in router for samf4

Added a new route in routes in frontend and fixed format in router.